my python tracer
CPython stock Lib/trace.py has inconvenient filter for path. It only
supports a blocklist of directories to ignore files that are under one
of them. It will be painful to get the behavior to only trace certain
set of files.

My tracer will accept a regular expression to filter the files that we
want to trace. Here is an example command to trace fx:

```
python mylib/mytracer.py main.py --path-filter="torch.fx"
```